\docType{data}
\name{databases}
\alias{databases}
\title{Database Data}
\format{A dataframe with ~700 observations on the following 5 variables,

\describe{

\item{Database}{String: Project Name}

\item{Slave}{String: Slave location}

\item{Lang_Code}{String: Project language}

\item{Project_Type}{String: Type, e.g. wiki}

\item{Lang_name}{String: Full language name}

}}
\description{
  Different datasets available based on past tool output
}
\keyword{datasets}

