\docType{data}
\name{gadgetdata}
\alias{gadgetdata}
\title{Gadget Data}
\format{A dataframe with ~5000 observations on the following 4 variables,

\describe{

\item{preference}{String: Name of the gadget}

\item{users}{Numeric: Number of users with that preference recorded}

\item{Project}{String: Project name}

\item{Project_Type}{String: Project type, e.g. wiki}

}}
\description{
  Gadget usage across projects and users.
}
\keyword{datasets}

