\name{MetaAnalysis-package}
\alias{MetaAnalysis-package}
\alias{MetaAnalysis}
\docType{package}
\title{
Meta-analysis
}
\description{
  This packages provides methods for meta-analysis of expression profiling data, including Fisher's combined probability test and meta-rank analysis.
}
\details{
\tabular{ll}{
Package: \tab MetaAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-27\cr
License: \tab What license is it under?\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Carolyn Ch'ng

Maintainer: Carolyn Ch'ng <cchng@chibi.ubc.ca>
%~~ The author and/or maintainer of the package ~~
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
